+++ CHANGELOG.md
@@ -64,6 +64,7 @@
 
 ## v1.5.x
 * Added: Daly BMS - Connect multiple BMS to the same RS485 port by @CaptKrisp
+* Changed: Rewritten code for external current sensor and fixed https://github.com/mr-manuel/venus-os_dbus-serialbattery/issues/60 by @mr-manuel
 
 ## v1.4.20240928